requirements.txt
backports-abc==0.4
certifi==2016.8.2
msgpack-python==0.4.8
msgpack-rpc-python==0.4
pika==0.10.0
sha3==0.2.1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.1
